chore: reorganize auth module folder
