+++ src/app/modules/auth/pages/register/register.component.spec.ts
